Add permission to get members of teams
We noticed that authorkind was incorrect for some PRs. For example, https://github.com/jupyterlab/jupyterlab/pull/17893 had authorkind of Seasoned Contributor, but the PR author has triage permissions. Adding the org member read-only permission solved the issue - the getCollaborators seems to now be returning the full list of people with triage/maintain/admin access.

However, I don't see exactly why it solved the issue. The list of collaborators returned before the permission update doesn't seem to follow a pattern. For example, each team with repo triage/maintain/admin access has people both in and not in the before list of collaborators.